Finished up the generators. It works now!
I also went through and spruced up formatting in all the files where it was needed. I want to add one more generator, the Point-Buy one, but it's a debate as to whether it has an actual spot within this or not. So while that's being decided, I want to add in a leveling ability (adds bonuses to the character abilities based on the class and level), then we can go from there.