Adding in Race Preference for Classes
Race Preference is a method of sorting races based on the bonuses they
grant, and how they affect being the chosen class. It is sorted by
preference in descending order. There are three classes of race when it
comes to any class, High preferrence, Average, and Poor preference, and
they are separated out as such.