Adding class preference for races.
Much like racial preferences for classes, if you want to pick a race first
(instead of a class), then it'll generate a list of preferred classes for
that race. This requires you to "lock" your choice however, or else it
will keep changing as you go back and forth between class and race. So we
added that ability too.